P3: build clean under -std=c23 / -std=c++23
The build now completes for both chips and all four formats (VST3, LV2, AAX,
Standalone), and both standalones launch with a working window.

Retired the forked chip emulators. sources/{opl3,opn2}/adl/measurer/chips was a
copy of libADLMIDI/libOPNMIDI 1.4's emulator code, 6,136 lines of it, and it
declared adl_audioTickHandler / opn2_audioTickHandler extern -- symbols the
current libraries no longer export, so it would have failed to link. measurer
only ever used setRate/writeReg/generate through the abstract base, and those
signatures are unchanged upstream, so it now drives the submodules' own
DosBoxOPL3 and MameOPN2 directly. The classes moved out of the ADL:: / OPN::
namespaces and MameOPN2 gained an OPNFamily constructor argument; nothing else
had to change. Note that the newer emulators are more accurate, so measured
instrument durations may shift slightly.

Language-level fixes:
  - M_PI -> std::numbers::pi_v<double>; the MSVC CRT does not define M_PI
    without _USE_MATH_DEFINES, and the old build assumed MinGW
  - NOMINMAX / WIN32_LEAN_AND_MEAN, so <windows.h> stops eating std::min
  - dropped 17 u8"" prefixes; since C++20 those are const char8_t[] and no
    longer initialise a const char* array
  - removed the `namespace ex { #include <juce_MidiKeyboardComponent.cpp> }`
    hack; juce_audio_utils is linked properly now
  - WOPLFile_Ptr / WOPNFile_Ptr moved into wopx_file.h, upstream having
    dropped them from the wopl/wopn headers

JUCE API removals:
  - Font::getStringWidthFloat -> GlyphArrangement::getStringWidth
  - Colour::fromHSV(..., 0xff) was ambiguous between the float and uint8
    alpha overloads

JUCE 8 removed the blocking modal calls unless JUCE_MODAL_LOOPS_PERMITTED is
set, which must stay off in a plugin -- a modal loop there freezes the host.
Every menu, file chooser and confirmation is now asynchronous: the code that
followed a blocking call moved into a continuation (finish_add_program,
finish_load_bank, finish_save_bank, finish_change_keymap, confirm_overwrite),
and select_emulator_by_menu reports through a callback instead of returning.
All continuations go through Component::SafePointer, since the editor can be
destroyed while a dialog is open. FileChooser is held in a member because
launchAsync returns immediately.

The language standard is set per-target rather than globally. CMAKE_C_STANDARD
and friends are inherited by add_subdirectory() dependencies, and a global
CMAKE_C_EXTENSIONS=OFF turned libADLMIDI's -std=gnu90 into -std=c90, which
breaks its `static inline` chip code. Our targets get strict -std=c23 /
-std=c++23; the submodules keep their own.

Not verified here: that the plugins actually produce correct audio, and the
AU/AAX behaviour (no macOS machine, no Pro Tools).

+++ sources/ui/components/vu_meter.cc
@@ -106,6 +106,6 @@ void Vu_Meter::update_gradient()
     for (unsigned s = 0; s < num_stops; ++s) {
         double r = s / (double)(num_stops - 1);
         double hue = hue_start + r * hue_range;
-        gradient.addColour(r, Colour::fromHSV(hue, 0.75, 0.75, 0xff));
+        gradient.addColour(r, Colour::fromHSV((float)hue, 0.75f, 0.75f, (uint8)0xff));
     }
 }
